% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_optimalMSAs.R
\name{align_part_set}
\alias{align_part_set}
\title{Co-optimal alignments from tree partitions}
\usage{
align_part_set(x, partition_set, exec, method, msa.program, type = type,
  coopt.sub = "all")
}
\description{
Co-optimal alignments from tree partitions
}
