% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residue_score_heatmap.R
\name{confidence.heatmap}
\alias{confidence.heatmap}
\title{Plot heatmap of cofidence on MSA}
\usage{
confidence.heatmap(obj, file = NULL, title, legend.text = TRUE,
  guidance_score = TRUE)
}
\arguments{
\item{obj}{output of guidance, guidance2 or HoT}

\item{file}{if a path is supplied the alignment is stored in a pdf, if file=NULL (default) alignment is promted}

\item{title}{title of the plot}

\item{legend.text}{logical, if TRUE legend is printed}

\item{guidance_score}{logical, if TRUE a barplot with the GUIDANCE score is printed}
}
\description{
Plots a heatmap of residue scores values on a MSA
}
\author{
Franz-Sebastian Krah
}
