% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Shortest Path}
\usage{
dijkstra(graph, initial_node)
}
\arguments{
\item{graph}{is A data frame}

\item{initial_node}{is a number for starting nod}
}
\value{
the shortest path between the selected nod and all other nods
}
\description{
Shortest Path
}
\examples{
dijkstra (wiki_graph , 3)
dijkstra (wiki_graph , 5)
}
\references{
(https://en.wikipedia.org/wiki/Dijkstra%27s algorithm)
}
