% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equalSubtrees.R
\name{equalSubtrees}
\alias{equalSubtrees}
\title{Split Phylogeny Equally}
\usage{
equalSubtrees(phy, k)
}
\arguments{
\item{phy}{A phylogenetic tree of class \code{\link{phylo}}.}

\item{k}{An integer, only trees with > k tips will be split.}
}
\value{
A list containing two subtrees of class \code{\link{"phylo}}.
}
\description{
Split a phylogenetic tree into two equally sized subtrees.
}
\details{
The phylogenetic tree is treated as unrooted, no matter whether is
  rooted or not.
}
\seealso{
\code{\link{decomposePhylo}} to decompose large phylogenetic trees
  into subsets of a certain size.
}
