% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Greatest common divisor}
\usage{
euclidean(num1, num2)
}
\arguments{
\item{num1}{is a number}

\item{num2}{is a number}
}
\value{
the Greatest common divisor between two numbers
}
\description{
Greatest common divisor
}
\references{
https://en.wikipedia.org/wiki/Euclidean algorithm
}
