% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{format_for_send}
\alias{format_for_send}
\title{Convert R objects to strings}
\usage{
format_for_send(obj)
}
\arguments{
\item{obj}{any object}
}
\description{
Prepare R objects for sending to postgresql
}
\details{
R objects that will be written to postgresql must be converted to
characters as all data is transferred to the server as text. The S3 method
\code{foramt_for_send} accomplishes this. It accepts any object and returns
a character representation.

You can define new conversions by supplying your own S3 override of
\code{format_for_send}.
}

