% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidance2.R
\name{guidance2}
\alias{guidance2}
\title{GUIDetree-based AligNment ConficencE 2}
\usage{
guidance2(sequences, msa.program = "mafft", exec, bootstrap = 100,
  n.part = "auto", col.cutoff = "auto", seq.cutoff = "auto",
  mask.cutoff = "auto", parallel = TRUE, ncore = "auto",
  method = "auto", alt.msas.file, n.coopt = "auto")
}
\arguments{
\item{sequences}{An object of class \code{\link{DNAbin}} or \code{\link{AAbin}}
containing unaligned sequences of DNA or amino acids.}

\item{msa.program}{A charcter string giving the name of the MSA program,
currelty one of c("mafft", "muscle", "clustalw2"); MAFFT is default}

\item{exec}{A character string giving the path to the executable of the
alignment program.}

\item{bootstrap}{An integer giving the number of perturbated MSAs.}

\item{col.cutoff}{numberic between 0 and 1; specifies a cutoff to remove unreliable columns below the cutoff; either user supplied or "auto" (0.73)}

\item{seq.cutoff}{numberic between 0 and 1; specifies a cutoff to remove unreliable sequences below the cutoff; either user supplied of "auto" (0.5)}

\item{mask.cutoff}{specific residues below a certain cutoff are masked ('N' for DNA, 'X' for AA); either user supplied of "auto" (0.5)}

\item{parallel}{logical, if TRUE, specify the number of cores}

\item{ncore}{number of cores}

\item{method}{further arguments passed to mafft, default is "auto"}

\item{mask}{specific residues below a certain cutoff are masked ('N' for DNA, 'X' for AA)}
}
\value{
alignment_score: is the GUIDANCE alignment score

GUIDANCE_residue_score

GUIDANCE_score: is the GUIDANCE column score

GUIDANCE_sequence_score

guidance_msa: is the base MSA removed from unreliable columns below
  cutoff

base_msa
}
\description{
MSA reliability assessment GUIDANCE2 (Sela et al. 2015)
}
\details{
Calculates column confidence (and other scors) by comparing alternative MSAs generated by the GUIDANCE with varying gap opening panelty and the HoT methodology. First 100 alternative MSAs (with BP guide trees) with varying gap opening panelty are produced, then for each n (default = 4) co-optimal alignments are produced using HoT. The basic comparison between the BP MSAs and a reference MSA is if column residue pairs are identically aligned in all alternative MSAs compared with the base MSA (see \code{compareMSAs}).
}
\references{
Felsenstein J. 1985. Confidence limits on phylogenies: an
  approach using the bootstrap. Evolution 39:783–791

Penn et al. (2010). An alignment confidence score capturing
  robustness to guide tree uncertainty. Molecular Biology and Evolution
  27:1759--1767

Sela et al. (2015). GUIDANCE2: accurate detection of unreliable alignment regions accounting for the uncertainty of multiple parameters. Nucleic acids research 43:W7--W14

G. Landan and D. Graur (2008). Local reliability measures from sets of co-optimal multiple sequencesuence alignments. Pacific Symposium on Biocomputin. 13:15--24
}
\seealso{
\code{\link{compareMSAs}}, \code{\link{guidance}}, \code{\link{HoT}}
}
\author{
Franz-Sebastian Krah

Christoph Heibl
}
