% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residue_score_heatmap.R
\name{heatmap.msa}
\alias{heatmap.msa}
\title{Plot GUIDANCE Residue Scores}
\usage{
heatmap.msa(obj, file = NULL)
}
\arguments{
\item{obj}{is the output object of guidance}

\item{file}{if a path is supplied, the alignment is stored in a pdf, if file=NULL (default) alignment is promted in R}
}
\description{
Plots a heatmap of GUIDANCE residue scores
}
\author{
Franz-Sebastian Krah
}
