% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEV-HoT.R
\name{HoT}
\alias{HoT}
\title{Heads or Tails alignment reliability}
\usage{
HoT(sequences, msa.program = "mafft", exec, type, n.coopt = "auto",
  col.cutoff = "auto", seq.cutoff = "auto", mask.cutoff = "auto",
  parallel = FALSE, ncore = "auto", method = "auto", plot_guide = TRUE,
  alt.msas.file)
}
\arguments{
\item{sequences}{An object of class \code{\link{DNAbin}} or \code{\link{AAbin}}
containing unaligned sequences of DNA or amino acids.
@param cutoff specifies a cutoff to remove unreliable columns below the cutoff}

\item{msa.program}{A charcter string giving the name of the MSA program,
one of c("mafft", "muscle", "clustalo", "clustalw2"); MAFFT is default}

\item{exec}{A character string giving the path to the executable of the
alignment program.}

\item{col.cutoff}{numberic between 0 and 1; specifies a cutoff to remove unreliable columns below the cutoff; either user supplied or "auto" (0.73)}

\item{seq.cutoff}{numberic between 0 and 1; specifies a cutoff to remove unreliable sequences below the cutoff; either user supplied of "auto" (0.5)}

\item{mask.cutoff}{specific residues below a certain cutoff are masked ('N' for DNA, 'X' for AA); either user supplied of "auto" (0.5)}

\item{parallel}{logical, if TRUE, specify the number of cores}

\item{ncore}{number of cores (default is 'auto')}

\item{method}{further arguments passed to MAFFT, default is "auto"}
}
\value{
list containing following scores and alignments:

mean_scores residue pair score and mean column score

column_score

residue_column_score column reliability

residue_pair_residue_score

residual_pair_sequence_pair_score

residual_pair_sequence_score

residue_pair_score

base_msa

HoT_msa is the base_MSA removed from unreliable residues/columns/sequences below cutoffs
}
\description{
MSA reliability assessment HoT (Landan and Graur 2008)
}
\details{
Calculates column reliability (and other scors) by comparing alternative
  MSAs generated by aligning guide tree partitions as described in Landan and Graur (2008). For details see \code{compareMSAs}. 8*(N-3) alternative MSAs are generated by default, where N is the number of sequences.
}
\references{
G. Landan and D. Graur (2008). Local reliability measures from sets of co-optimal multiple sequencesuence alignments. 13:15--24
}
\seealso{
\code{\link{compareMSAs}}, \code{\link{guidance}}, \code{\link{guidance2}}
}
\author{
Franz-Sebastian Krah

Christoph Heibl
}
