% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEV-HoT.R
\name{HoT_dev}
\alias{HoT_dev}
\title{Heads or Tails alignment reliability}
\usage{
HoT_dev(sequences, msa.program = "mafft", exec, n.part = "auto",
  col.cutoff = "auto", seq.cutoff = "auto", mask.cutoff = "auto",
  parallel = FALSE, ncore, method = "auto", plot_guide = TRUE)
}
\arguments{
\item{sequences}{An object of class \code{\link{DNAbin}} or \code{\link{AAbin}}
containing unaligned sequencesuences of DNA or amino acids.#' @param cutoff specifies a cutoff to remove unreliable columns below the cutoff}

\item{msa.program}{A charcter string giving the name of the MSA program,
one of c("mafft", "muscle", "clustalo", "clustalw2"); MAFFT is default}

\item{exec}{A character string giving the path to the executable of the
alignment program.}

\item{n.part}{number of co-optimal alignments should be used (must be integer < Ntip - 3)}

\item{col.cutoff}{numberic between 0 and 1; specifies a cutoff to remove unreliable columns below the cutoff; either user supplied or "auto" (0.73)}

\item{seq.cutoff}{numberic between 0 and 1; specifies a cutoff to remove unreliable sequences below the cutoff; either user supplied of "auto" (0.5)}

\item{mask.cutoff}{specific residues below a certain cutoff are masked ('N' for DNA, 'X' for AA); either user supplied of "auto" (0.5)}

\item{parallel}{logical, if TRUE, specify the number of cores}

\item{ncore}{number of cores (default is 'auto')}

\item{method}{further arguments passed to MAFFT, default is "auto"}
}
\value{
alignment_reliability

residue_reliability

sequencesuence_reliability

column_reliability

HoT_MSA: is the base MSA removed from unreliable sites and/or sequences and/or masked residues below the cutoffs

base_msa
}
\description{
Assess reliability in an aligment using the
  HoT algorithm
}
\details{
Calculates column reliability by comparing alternative
  MSAs generated by aligning guide tree partitions.
}
\references{
G. Landan and D. Graur (2008). Local reliability measures from sets of co-optimal multiple sequencesuence alignments. 13:15--24
}
\author{
Franz-Sebastian Krah

Franz-Sebastian Krah

Christoph Heibl
}
