% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPtree_from_MSA.R
\name{msaBP_nj_tree}
\alias{msaBP_nj_tree}
\title{Generate NJ BP tree from MSA}
\usage{
msaBP_nj_tree(msa, outgroup)
}
\arguments{
\item{msa}{multiple sequence alignment of class AAbin or DNAbin}
}
\value{
tree of class phylo
}
\description{
Generate NJ BP tree from MSA
}
