% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasta.R
\name{pasta}
\alias{pasta}
\title{Ultra-Large Multiple Sequence Alignment with PASTA}
\usage{
pasta(seqs, gt, k = 200, cutoff = 0.93, parallel = FALSE,
  bootstrap = 100, msa.program = "mafft", method = "auto", exec, ncore)
}
\arguments{
\item{seqs}{An object of class \code{\link{DNAbin}} or \code{\link{AAbin}}
containing unaligned sequences of DNA or amino acids.}

\item{gt}{\emph{Currently unused.}}

\item{k}{An integer given the size of cluster in which the dataset is split.}
}
\description{
Provides a complete reimplementation of the PASTA algorithm (Mirarab, Nguyen, and Warnow 2014) in R.
}
