% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{set_path}
\alias{append_path}
\alias{get_path}
\alias{path_contains}
\alias{prepend_path}
\alias{set_path}
\title{PostgreSQL path variable}
\usage{
set_path(...)

get_path()

append_path(...)

prepend_path(...)

path_contains(x)
}
\arguments{
\item{...}{path names}
}
\description{
Manipulate the PostgreSQL path variable
}

