% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanningTree.R
\name{spanningTree}
\alias{spanningTree}
\title{Minimum-Spanning-Tree for PASTA}
\usage{
spanningTree(phy, subtrees)
}
\arguments{
\item{phy}{An object of class \code{\link{phylo}} containing a complete
phylogenetic tree.}

\item{subtrees}{A list containing object of class \code{\link{phylo}} as
obtained with \code{\link{centroidDecomposition}}.}
}
\description{
Creates miminum spanning tree connecting subsets obtained by
  centroid decomposition.
}
