% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki graph data}
\format{An object of class \code{data.frame} with 18 rows and 3 columns.}
\usage{
graphs
}
\description{
data for lab3
}
\references{
https://en.wikipedia.org/wiki/Graph
}
\keyword{datasets}
