% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_scores.R
\name{calc_scores}
\alias{calc_scores}
\title{Wrapper Function of C Code for Calculation of GUIDANCE Scores}
\usage{
calc_scores(ref, com, n_id = 1)
}
\arguments{
\item{ref}{of class data.frame, is the reference MSA ('BASE MSA') with sequences as columns}

\item{com}{like ref, but 1 alternative ('perturbated MSA')}

\item{n_id}{is used as 'i' if function is run in parallel}
}
\value{
list containing GUIDANCE column score, the residue score and sequence score
}
\description{
Wrapper Function of C Code for Calculation of GUIDANCE Scores
}
\author{
Franz-Sebastian Krah
}
