% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidance.R
\name{guidance}
\alias{guidance}
\title{Alignment Confidence Scores with GUIDANCE}
\usage{
guidance(seqs, cutoff = 0.93, parallel = FALSE, ncore, bootstrap = 100,
  msa.program = "mafft", exec, method = "auto")
}
\arguments{
\item{seqs}{An object of class \code{\link{DNAbin}} or \code{\link{AAbin}}
containing unaligned sequences of DNA or amino acids.}

\item{cutoff}{specifies a cutoff to remove unreliable columns below the
cutoff}

\item{parallel}{logical, if TRUE, specify the number of cores}

\item{ncore}{number of cores}

\item{bootstrap}{An integer giving the number of perturbated MSAs.}

\item{msa.program}{A charcter string giving the name of the MSA program,
currelty either \code{"mafft"} (default) or\code{"prank"}.}

\item{exec}{A character string giving the path to the executable of the
alignment program.}

\item{method}{further arguments passed to mafft, default is "auto"}
}
\value{
alignment_score: is the GUIDANCE alignment score

GUIDANCE_residue_score

GUIDANCE_score: is the GUIDANCE column score

GUIDANCE_sequence_score

guidance_msa: is the base MSA removed from unreliable columns below
  cutoff

base_msa
}
\description{
Assess uncertainty in multiple sequence aligment (MSA) with the
  GUIDANCE algorithm (Penn et al. 2010).
}
\details{
Calculates column unreliability by comparing alternative bootstrap
  MSAs (Felsenstein 1985) generated by different guide trees. Thus GUIDANCE
  is a measure for guide-tree uncertainty. The basic comparison between the
  BP MSAs and a reference MSA is if column residue pairs are identical in
  both MSAs.
}
\references{
Felsenstein J. 1985. Confidence limits on phylogenies: an
  approach using the bootstrap. Evolution 39:783–791

Penn et al. (2010). An alignment confidence score capturing
  robustness to guide tree uncertainty. Molecular Biology and Evolution
  27:1759--1767
}
\author{
Franz-Sebastian Krah
}
