% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidance.R
\name{guidance}
\alias{guidance}
\title{GUIDetree-based AligNment ConficencE}
\usage{
guidance(sequences, msa.program = "mafft", exec, bootstrap = 100,
  col.cutoff = "auto", seq.cutoff = "auto", mask.cutoff = "auto",
  parallel = FALSE, ncore = "auto", method = "auto", alt.msas.file)
}
\arguments{
\item{msa.program}{A charcter string giving the name of the MSA program,
currelty one of c("mafft", "muscle", "clustalo", "clustalw2"); MAFFT is default}

\item{exec}{A character string giving the path to the executable of the
alignment program.}

\item{bootstrap}{An integer giving the number of perturbated MSAs.}

\item{col.cutoff}{numberic between 0 and 1; specifies a cutoff to remove unreliable columns below the cutoff; either user supplied or "auto" (0.73)}

\item{seq.cutoff}{numberic between 0 and 1; specifies a cutoff to remove unreliable sequences below the cutoff; either user supplied of "auto" (0.5)}

\item{mask.cutoff}{specific residues below a certain cutoff are masked ('N' for DNA, 'X' for AA); either user supplied of "auto" (0.5)}

\item{parallel}{logical, if TRUE, specify the number of cores}

\item{ncore}{number of cores (default is maxinum of local threads)}

\item{method}{further arguments passed to mafft, default is "auto"}

\item{sequencess}{An object of class \code{\link{DNAbin}} or \code{\link{AAbin}}
containing unaligned sequences of DNA or amino acids.}
}
\value{
alignment_score: is the GUIDANCE alignment score

GUIDANCE_residue_score

GUIDANCE_score: is the GUIDANCE column score

GUIDANCE_sequence_score

guidance_msa: is the base MSA removed from unreliable columns below
  cutoff

base_msa
}
\description{
Assess uncertainty in multiple sequence aligment (MSA) with the
  GUIDANCE algorithm (Penn et al. 2010).
}
\details{
Calculates column confidence by comparing alternative MSAS generated by bootstrap MSAs (Felsenstein 1985) and derived guide trees. The basic comparison between the BP MSAs and a reference MSA is if column residue pairs are identical in both MSAs.
}
\references{
Felsenstein J. 1985. Confidence limits on phylogenies: an
  approach using the bootstrap. Evolution 39:783–791

Penn et al. (2010). An alignment confidence score capturing
  robustness to guide tree uncertainty. Molecular Biology and Evolution
  27:1759--1767
}
\author{
Franz-Sebastian Krah

Franz-Sebastian Krah

Christoph Heibl
}
