% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidance2-DEV.R
\name{guidance2}
\alias{guidance2}
\title{Alignment Confidence Scores with GUIDANCE}
\usage{
guidance2(sequences, msa.program = "mafft", exec, bootstrap = 100,
  n.part = "auto", col.cutoff = "auto", seq.cutoff = "auto",
  mask.cutoff = "auto", parallel = TRUE, ncore = "auto",
  method = "auto", alt.msas.file, n.coopt = "auto")
}
\arguments{
\item{msa.program}{A charcter string giving the name of the MSA program,
currelty one of c("mafft", "muscle", "clustalw2"); MAFFT is default}

\item{exec}{A character string giving the path to the executable of the
alignment program.}

\item{bootstrap}{An integer giving the number of perturbated MSAs.}

\item{parallel}{logical, if TRUE, specify the number of cores}

\item{ncore}{number of cores}

\item{method}{further arguments passed to mafft, default is "auto"}

\item{seq}{is the raw input sequences of class DNAbin (DNA) or SeqFastaAA (AA)}

\item{cutoff}{specifies a cutoff to remove unreliable columns below the cutoff}

\item{seqs}{An object of class \code{\link{DNAbin}} or \code{\link{AAbin}}
containing unaligned sequences of DNA or amino acids.}

\item{cutoff}{specifies a cutoff to remove unreliable columns below the
cutoff}

\item{mask}{specific residues below a certain cutoff are masked ('N' for DNA, 'X' for AA)}
}
\value{
alignment_score: is the GUIDANCE alignment score

GUIDANCE_residue_score

GUIDANCE_score: is the GUIDANCE column score

GUIDANCE_sequence_score

guidance_msa: is the base MSA removed from unreliable columns below
  cutoff

base_msa
}
\description{
Assess uncertainty in multiple sequence aligment (MSA) with the
  GUIDANCE2 algorithm based on GUIDANCE (Penn et al. 2010) and HoT (G. Landan and D. Graur 2008)
}
\details{
GUIDetree-based AligNment ConficencE 2

Calculates column unreliability by comparing alternative bootstrap
  MSAs (Felsenstein 1985) generated by different guide trees. Thus GUIDANCE
  is a measure for guide-tree uncertainty. The basic comparison between the
  BP MSAs and a reference MSA is if column residue pairs are identical in
  both MSAs.
}
\references{
Felsenstein J. 1985. Confidence limits on phylogenies: an
  approach using the bootstrap. Evolution 39:783–791

Penn et al. (2010). An alignment confidence score capturing
  robustness to guide tree uncertainty. Molecular Biology and Evolution
  27:1759--1767

Sela et al. (2015). GUIDANCE2: accurate detection of unreliable alignment regions accounting for the uncertainty of multiple parameters. Nucleic acids research 43:W7--W14

G. Landan and D. Graur (2008). Local reliability measures from sets of co-optimal multiple sequencesuence alignments. Pacific Symposium on Biocomputin. 13:15--24
}
\author{
Franz-Sebastian Krah

Franz-Sebastian Krah

Christoph Heibl
}
