% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residue_score_heatmap.R
\name{guidance_heatmap}
\alias{guidance_heatmap}
\title{Plot GUIDANCE Residue Scores}
\usage{
guidance_heatmap(guidance_object, file = NULL)
}
\arguments{
\item{guidance_object}{is simply the output object of guidance}

\item{file}{if a dir path is supplied, the alignment is stored in a pdf, if file=NULL (default) alignment is plotted in R}
}
\description{
Plots a heatmap of GUIDANCE residue scores
}
\author{
Franz-Sebastian Krah
}
