% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoT.R
\name{HoT}
\alias{HoT}
\title{Heads or Tails alignment reliability}
\usage{
HoT(sequences, cutoff = 0.93, parallel = FALSE, ncore,
  msa.program = "mafft", method = "auto", mask = FALSE, mafft_exec,
  plot_guide = TRUE)
}
\arguments{
\item{sequences}{An object of class \code{\link{DNAbin}} or \code{\link{AAbin}}
containing unaligned sequencesuences of DNA or amino acids.#' @param cutoff specifies a cutoff to remove unreliable columns below the cutoff}

\item{cutoff}{specifies a cutoff to remove unreliable columns below the
cutoff}

\item{parallel}{logical, if TRUE, specify the number of cores}

\item{ncore}{number of cores}

\item{msa.program}{A charcter string giving the name of the MSA program,
currelty either \code{"mafft"} (default) or\code{"prank"}.}

\item{method}{further arguments passed to MAFFT, default is "auto"}

\item{mask}{specific residues below a certain cutoff are masked ('N' for DNA, 'X' for AA)}

\item{mafft_exec}{A character string giving the path to the executable of MAFFT
(usually: '/usr/local/bin/mafft')}
}
\value{
alignment_reliability

residue_reliability

sequencesuence_reliability

column_reliability

HoT_MSA: is the base MSA removed from unreliable columns below
  \code{cutoff}

base_msa
}
\description{
Assess reliability in an aligment using the
  HoT algorithm
}
\details{
Calculates column reliability by comparing alternative
  MSAs generated by aligning guide tree partitions.
}
\references{
G. Landan and D. Graur (2008). Local reliability measures from sets of co-optimal multiple sequencesuence alignments. 13:15--24
}
\author{
Franz-Sebastian Krah

Franz-Sebastian Krah

Christoph Heibl
}
