% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rpg.R, R/utils.R
\name{libpq_version}
\alias{createdb}
\alias{dropdb}
\alias{enable_extension}
\alias{enable_postgis}
\alias{encrypt_password}
\alias{get_encoding}
\alias{libpq_version}
\alias{set_encoding}
\alias{set_error_verbosity}
\alias{toggle_echo}
\title{Miscelaneous functions}
\usage{
libpq_version()

encrypt_password(passwd, user)

get_encoding()

set_encoding(encoding)

set_error_verbosity(verbosity)

toggle_echo()

enable_postgis(schemaname = "postgis")

enable_extension(extension, schemaname = extension)

createdb(name, opts = "", description = "")

dropdb(name, opts = "")
}
\arguments{
\item{passwd}{the password}

\item{user}{the user name}

\item{encoding}{the character encoding}

\item{verbosity}{one of "terse", "default", "verbose"}

\item{schemaname}{install in this schema}

\item{extension}{the extension name}

\item{name}{name of the database}

\item{opts}{a list of command options}

\item{description}{a description string}
}
\description{
Various utility functions
}
\details{
\code{enable_postgis} will attempt to install the postgis
extension in the named schema using \code{enable_extension}.

\code{enable_extension} will attempt to install the
extension in the named schema. The default search path is altered to
include the new schema.
}
\examples{
\dontrun{
# try connecting to default database
createdb("rpgtesting")
connect("rpgtesting")
begin()

libpq_version()
encrypt_password("test", "tester")
get_encoding()
set_encoding("UTF8")
set_error_verbosity("terse")
set_error_verbosity("verbose")
set_error_verbosity("default")
enable_postgis()

# cleanup
rollback()
disconnect()
dropdb("rpgtesting")}

}
\author{
Timothy H. Keitt
}

