% Generated by roxygen2 (4.0.1): do not edit by hand
\name{libpq_version}
\alias{encrypt_password}
\alias{get_encoding}
\alias{libpq_version}
\alias{set_encoding}
\alias{set_error_verbosity}
\title{Miscelaneous functions}
\usage{
libpq_version()

encrypt_password(passwd, user)

get_encoding()

set_encoding(encoding)

set_error_verbosity(verbosity)
}
\arguments{
\item{passwd}{the password}

\item{user}{the user name}

\item{encoding}{the character encoding}

\item{verbosity}{one of "terse", "default", "verbose"}
}
\description{
Various utility functions
}
\examples{
\dontrun{
libpq_version()
encrypt_password("test", "tester")
get_encoding()
set_encoding("UTF8")
set_error_verbosity("terse")
set_error_verbosity("verbose")
set_error_verbosity("default")}
}
\author{
Timothy H. Keitt
}

