% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sql.R
\name{create_index}
\alias{create_index}
\title{Create an index}
\usage{
create_index(tablename, columnname, unique = FALSE, concurrently = FALSE,
  if_not_exists = FALSE, using = NULL, schemaname = NULL)
}
\arguments{
\item{tablename}{the name of the table}

\item{schemaname}{specifically in this schema}

\item{columname}{the name of the column}
}
\description{
Create and index on an existing table column
}
\details{
This function is not yet finished
}

