% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sql.R
\name{create_index}
\alias{create_index}
\title{Create an index}
\usage{
create_index(tablename, columnname, unique = FALSE, using = NULL,
  schemaname = NULL)
}
\arguments{
\item{tablename}{the name of the table}

\item{columnname}{the name of the column}

\item{unique}{if true, create a unique index}

\item{using}{the index method}

\item{schemaname}{specifically in this schema}
}
\description{
Create and index on an existing table column
}
\details{
This function is not yet finished
}

